content/base/test/test_bug413974.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=413974
-->
<head>
  <title>Test for Bug 413974</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=413974">Mozilla Bug 413974</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 413974 **/
var req = new XMLHttpRequest();
req.open("POST", window.location.href);
req.setRequestHeader("Content-Type", "text/plain; boundary=01234567890");
req.send("Some text");

is(SpecialPowers.wrap(req).channel
      .QueryInterface(SpecialPowers.Ci.nsIHttpChannel)
      .getRequestHeader("Content-Type"),
   "text/plain; charset=UTF-8; boundary=01234567890",
   "Charset should come before boundary");
</script>
</pre>
</body>
</html>
